Revert "Resolves #1"
This reverts commit d022d9407bf637288b5c90772842ac663291c5dd.
