+++ src/main/kotlin/setu/controllers/HealthTrackerController.kt
@@ -1,0 +1,96 @@
+package setu.controllers
+
+import com.fasterxml.jackson.databind.SerializationFeature
+import io.javalin.http.Context
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import setu.domain.Activity
+import setu.domain.User
+import setu.domain.repository.ActivityDao
+import setu.domain.repository.UserDao
+import com.fasterxml.jackson.datatype.joda.JodaModule
+
+object HealthTrackerController {
+
+    private val userDao = UserDao()
+    private val activityDao = ActivityDao()
+
+    fun getAllUsers(ctx: Context) {
+        ctx.json(userDao.getAll())
+    }
+
+    fun getUserByUserId(ctx: Context) {
+        val user = userDao.findById(ctx.pathParam("user-id").toInt())
+        if (user != null) {
+            ctx.json(user)
+        }
+    }
+
+    fun addUser(ctx: Context) {
+        val mapper = jacksonObjectMapper()
+        val user = mapper.readValue<User>(ctx.body())
+        userDao.save(user)
+        ctx.json(user)
+    }
+
+    fun getUserByEmail(ctx: Context) {
+        val user = userDao.findByEmail(ctx.pathParam("email"))
+        if (user != null) {
+            ctx.json(user)
+        }
+    }
+
+    fun deleteUserByUserId(ctx: Context) {
+        userDao.delete(ctx.pathParam("user-id").toInt())
+    }
+
+    fun updateUser(ctx: Context) {
+        val mapper = jacksonObjectMapper()
+        val user = mapper.readValue<User>(ctx.body())
+        userDao.update(
+            id = ctx.pathParam("user-id").toInt(),
+            user = user
+        )
+    }
+
+    //--------------------------------------------------------------
+    // ActivityDAOI specifics
+    //-------------------------------------------------------------
+
+    fun getAllActivities(ctx: Context) {
+        //mapper handles the deserialization of Joda date into a String.
+        val mapper = jacksonObjectMapper()
+            .registerModule(JodaModule())
+            .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+        ctx.json(mapper.writeValueAsString(activityDao.getAll()))
+    }
+
+    fun getActivitiesByUserId(ctx: Context) {
+        if (userDao.findById(ctx.pathParam("user-id").toInt()) != null) {
+            val activities = activityDao.findByUserId(ctx.pathParam("user-id").toInt())
+            if (activities != null) {
+                //mapper handles the deserialization of Joda date into a String.
+                val mapper = jacksonObjectMapper()
+                    .registerModule(JodaModule())
+                    .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+                ctx.json(mapper.writeValueAsString(activities))
+            }
+        }
+    }
+
+    fun addActivity(ctx: Context) {
+        //mapper handles the serialisation of Joda date into a String.
+        val mapper = jacksonObjectMapper()
+            .registerModule(JodaModule())
+            .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+        val activity = mapper.readValue<Activity>(ctx.body())
+        activityDao.save(activity)
+        ctx.json(activity)
+    }
+
+    fun deleteActivitiesByUserId(ctx: Context) {}
+    fun getActivityById(ctx: Context) {}
+    fun deleteActivityById(ctx: Context) {}
+    fun updateActivityById(ctx: Context) {}
+
+}
